refactor: extract common file processing
