+++ src/Commands/Refactor/VarCommentsRefactorer.php
@@ -8,6 +8,7 @@ use Symfony\Component\Console\Command\Command;
 use Vix\Syntra\Commands\SyntraRefactorCommand;
 use Vix\Syntra\Enums\ProgressIndicatorType;
 use Vix\Syntra\Facades\File;
+use Vix\Syntra\Traits\ProcessesFilesTrait;
 
 /**
  * Refactors @var comments by standardizing the order of type and variable name.
@@ -17,6 +18,7 @@ use Vix\Syntra\Facades\File;
  */
 class VarCommentsRefactorer extends SyntraRefactorCommand
 {
+    use ProcessesFilesTrait;
     protected ProgressIndicatorType $progressType = ProgressIndicatorType::PROGRESS_BAR;
 
     protected function configure(): void
@@ -30,40 +32,9 @@ class VarCommentsRefactorer extends SyntraRefactorCommand
 
     public function perform(): int
     {
-        $files = File::collectFiles($this->path);
-
-        $this->setProgressMax(count($files));
-        $this->startProgress();
-
-        foreach ($files as $filePath) {
-            $content = file_get_contents($filePath);
-
-            $newContent = $this->convertVarComments($content);
-
-            if (!$this->dryRun) {
-                File::writeChanges($filePath, $content, $newContent);
-            }
-
-            $this->advanceProgress();
-        }
-
-        $this->finishProgress();
-
-        $changed = File::getChangedFiles();
-        File::clearChangedFiles();
-
-        if ($changed) {
-            $this->output->section('Changed files');
-            $list = array_map(
-                fn (string $f): string => File::makeRelative($f, $this->path),
-                $changed
-            );
-            $this->listing($list);
-        } else {
-            $this->output->success('No files needed updating.');
-        }
-
-        return Command::SUCCESS;
+        return $this->processFiles(function (string $content): string {
+            return $this->convertVarComments($content);
+        });
     }
 
     /**
